Improve Time Factor Handling
As before, there are two modes: "(scaled) realtime approximation" and
"as fast as possible". The main change now is to introduce a feedback
loop that adjusts the sleep time and the scheduled number of simulated
MCU clock ticks per "control frame", while trying to keep the duration
of this control frame constant. This fixes two issues:
 - The previous code suffered from the "spiral of death", where each
   frame got longer and longer if the host was too slow: to catch up,
   even more ticks got scheduled, taking even more time to complete.
 - There was a very small and fixed amount of sleep, even in "fast mode",
   that according to the original comment should yield the control back
   to the OS to allow the other threads to get scheduled. With a modern
   preemptive OS this should not be necessary -- we now simply trust the
   OS' scheduler.
   This massively reduces the time spend in kernel-mode and is in fact a
   precondition for the "fast mode" to be fast: Previously there has been
   one sleep, and thus one potential context switch, per simulated tick,
   resulting in "fast mode" being utterly slow.

Then, the `speedup` in the simulator state structure is now the current
speedup, calculated over the last control frame. The desired target value
given by the user via CLI flag, is moved to the `args` structure.
The current speedup is now available in both modes, i.e. it's also for
the "fast mode".