Setup Github Actions for Build and Test
Based on the cmake-multi-platform template [1] a first CI pipelie is set
up, that builds on Windows, Linux and macOS, using gcc and clang.
As rudimentary smoke tests the resulting executables are started with `-h`.

Windows required a bit of extra work as the default compiler, Microsoft's
cl.exe cannot compile grbl, so gcc from MinGW64 is used. But this one does
not work with the default CMake generator on Windows, which is MSBuild.

[1]: https://github.com/actions/starter-workflows/blob/main/ci/cmake-multi-platform.yml